requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==3.2.5
djangorestframework==3.12.4
faiss-cpu==1.8.0
filelock==3.15.3
Flask==3.0.3
fsspec==2024.6.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==2.0.0
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pybind11==2.12.0
pytz==2024.1
setuptools==70.1.0
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.0.3
wheel==0.43.0
